backend/pyproject.toml
[project]
name = "quantumstock-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.135.1",
    "ollama>=0.4.0",
    "pdfplumber>=0.11.9",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "scrapy>=2.14.1",
    "spacy>=3.7.0",
    "uvicorn[standard]>=0.41.0",
    "yfinance>=1.2.0",
]
